requirements.txt

matplotlib==3.1.1
numpy==1.17.2
pandas==0.25.1
pytest==5.1.3
scikit-learn==0.21.3
scikit-surprise==1.1.0
scipy==1.3.1
six==1.12.0
Flask==1.1.1
